// src/app/types/enums.ts
export enum Chart {
  BAR_CHART = "Bar Chart",
  LINE_CHART = "Line Chart",
  PIE_CHART = "Pie Chart",
}

export enum Categories {
  Featured = "Featured",
  KPI = "KPI",
  Layouts = "Layouts",
  Storyboards = "Storyboards",
}

export enum AssetType {
  KPI = "KPI",
  Layouts = "Layout",
  Storyboards = "Storyboard",
}